Rename addwidget package to onecontactwidget
