+++ feature/onecontactwidget/src/main/kotlin/dev/ferp/dcw/feature/onecontactwidget/WidgetListScreen.kt
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package dev.ferp.dcw.feature.addwidget
+package dev.ferp.dcw.feature.onecontactwidget
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
